Added RBIG TC implementation with scan.
