+++ rbig_jax/transforms/linear.py
@@ -42,6 +42,37 @@ def compute_projection(X: np.ndarray) -> np.ndarray:
     # Compute SVD
     _, _, VT = np.linalg.svd(X, full_matrices=False, compute_uv=True)
 
+    return X @ VT.T
+
+
+def compute_projection_v1(X: np.ndarray) -> np.ndarray:
+    """Compute PCA projection matrix
+    Using SVD, this computes the PCA components for
+    a dataset X and computes the projection matrix
+    needed to do the PCA decomposition.
+
+    Parameters
+    ----------
+    X : np.ndarray, (n_samples, n_features)
+        the data to calculate to PCA projection matrix
+    
+    Returns
+    -------
+    VT : np.ndarray, (n_features, n_features)
+        the projection matrix (V.T) for the PCA decomposition
+
+    Notes
+    -----
+    Can find the original implementation here:
+    https://bit.ly/2EBDV9o
+    """
+
+    # center the data
+    X = X - np.mean(X, axis=0)
+
+    # Compute SVD
+    _, _, VT = np.linalg.svd(X, full_matrices=False, compute_uv=True)
+
     return VT.T
